+++ tests/factories/TestFactories.php
@@ -1,0 +1,7 @@
+<?php
+$factory->define(Post::class, function ($faker) {
+    return [
+        'title' => $faker->sentence(5),
+        'body'  => $faker->paragraph,
+    ];
+});